parse and distinguish locations and movements
Now locations are geographical references and a location
can have one or more movements.

When starting a video by alternating a hover between two
locations and one of the locations has multiple movements
(such as "Newmark St & 16th St") the movement played will
step through the movements associated with the location
in sequence.

    {
      "address": "Newmark St & 16th St",
      "latitude": 42.32679199977638,
      "longitude": -83.07759100024275,
      "movements": [
        11,
        144
      ]
    },