sequence through waltz movements in any order
Now id a waltz is started in movement "B" it will progress
through movements "C" and "A" before it is done.

Also scheduling interviews is now done by map page by
extending waltzLocation structure with new eventType and
eventData keys.

The video page interprets three different eventTypes now:

1. "testing"

A boolean indicating whether app is running in testing (debug)
mode or not.

2. "movement"

The data in waltzLocation indicates a movement that should be
played.

3. "interview"

The data in waltzLocation indicates an interview that should be
played.