disable full-screen exit on click/touch in map/video
When not in full-screen mode there is a button (four
small arrows) in the upper right of the content area
of the map and video pages.

This button is a "toggle" button so when in full-screen
mode this button is invisible but still there. Clicking
on it will toggle the page back out of full-screen mode.

When pages are in testing mode clicking invisible button when
in full-screen mode will exit full-screen mode.

However now when testing mode is turned off clicking/touching
the invisible button has no effect.

User must use keyboard to exit full-screen mode.